+++ process_manager/views/__init__.py
@@ -1,0 +1,1 @@
+"""Module for app view functions."""
